+++ changelog.md
@@ -109,6 +109,9 @@ with other backends. see #9125. Use `-d:nimLegacyJsRound` for previous behavior.
 - Added `math.signbit`.
 
 - Removed the optional `longestMatch` parameter of the `critbits._WithPrefix` iterators (it never worked reliably)
+- In `lists`: renamed `append` to `add` and added a deprecated `append`;
+  added `prepend` and `prependMoved` analogously to `add` and `andMoved`;
+  added `remove` for `SinglyLinkedList`s.
 
 ## Language changes